(logging): this introduces a fs2 based task logging implementation (#29)
* (logging): this introduces a fs2 based task logging implementation and
changes the corresponding base trait. this makes it possible to use
fs2-blobstore to integrate with AWS S3 and also makes chunking
and general control of the log stream easier. The non chunking file
logger is still available if needed. The command is now running in
parallel and the process logger will use a fs2 queue to publish new
lines.