(execution): use cats IO in apply schedule
The old implementation of was hard to reason about in terms of "what
happens when" since we start several futures in parallel and sequence
them. With IO we can now reason better about about that sequencing and
hence the resources (mostly DB) used for applying.

Setting of the due date was also ignoring errors, which is now not the
case.

This also removes the unsafe recursion in the cron scheduler, replacing
it with a stream unfold.